+++ src/Symfony/Component/Ldap/Adapter/ConnectionInterface.php
@@ -11,6 +11,10 @@
 
 namespace Symfony\Component\Ldap\Adapter;
 
+use Symfony\Component\Ldap\Exception\AlreadyExistsException;
+use Symfony\Component\Ldap\Exception\ConnectionTimeoutException;
+use Symfony\Component\Ldap\Exception\InvalidCredentialsException;
+
 /**
  * @author Charles Sarrazin <charles@sarraz.in>
  */
@@ -25,6 +29,10 @@ interface ConnectionInterface
 
     /**
      * Binds the connection against a user's DN and password.
+     *
+     * @throws AlreadyExistsException      When the connection can't be created because of an LDAP_ALREADY_EXISTS error
+     * @throws ConnectionTimeoutException  When the connection can't be created because of an LDAP_TIMEOUT error
+     * @throws InvalidCredentialsException When the connection can't be created because of an LDAP_INVALID_CREDENTIALS error
      */
     public function bind(string $dn = null, string $password = null);
 }