+++ src/Symfony/Component/Messenger/Transport/RedisExt/Connection.php
@@ -31,6 +31,7 @@ class Connection
         'group' => 'symfony',
         'consumer' => 'consumer',
         'auto_setup' => true,
+        'stream_max_entries' => 0, // any value higher than 0 defines an approximate maximum number of stream entries
     ];
 
     private $connection;
@@ -38,6 +39,7 @@ class Connection
     private $group;
     private $consumer;
     private $autoSetup;
+    private $maxEntries;
     private $couldHavePendingMessages = true;
 
     public function __construct(array $configuration, array $connectionCredentials = [], array $redisOptions = [], \Redis $redis = null)
@@ -49,10 +51,16 @@ class Connection
         $this->connection = $redis ?: new \Redis();
         $this->connection->connect($connectionCredentials['host'] ?? '127.0.0.1', $connectionCredentials['port'] ?? 6379);
         $this->connection->setOption(\Redis::OPT_SERIALIZER, $redisOptions['serializer'] ?? \Redis::SERIALIZER_PHP);
+
+        if (isset($connectionCredentials['auth'])) {
+            $this->connection->auth($connectionCredentials['auth']);
+        }
+
         $this->stream = $configuration['stream'] ?? self::DEFAULT_OPTIONS['stream'];
         $this->group = $configuration['group'] ?? self::DEFAULT_OPTIONS['group'];
         $this->consumer = $configuration['consumer'] ?? self::DEFAULT_OPTIONS['consumer'];
         $this->autoSetup = $configuration['auto_setup'] ?? self::DEFAULT_OPTIONS['auto_setup'];
+        $this->maxEntries = $configuration['stream_max_entries'] ?? self::DEFAULT_OPTIONS['stream_max_entries'];
     }
 
     public static function fromDsn(string $dsn, array $redisOptions = [], \Redis $redis = null): self
@@ -70,6 +78,7 @@ class Connection
         $connectionCredentials = [
             'host' => $parsedUrl['host'] ?? '127.0.0.1',
             'port' => $parsedUrl['port'] ?? 6379,
+            'auth' => $parsedUrl['pass'] ?? $parsedUrl['user'] ?? null,
         ];
 
         if (isset($parsedUrl['query'])) {
@@ -82,7 +91,19 @@ class Connection
             unset($redisOptions['auto_setup']);
         }
 
-        return new self(['stream' => $stream, 'group' => $group, 'consumer' => $consumer, 'auto_setup' => $autoSetup], $connectionCredentials, $redisOptions, $redis);
+        $maxEntries = null;
+        if (\array_key_exists('stream_max_entries', $redisOptions)) {
+            $maxEntries = filter_var($redisOptions['stream_max_entries'], FILTER_VALIDATE_INT);
+            unset($redisOptions['stream_max_entries']);
+        }
+
+        return new self([
+            'stream' => $stream,
+            'group' => $group,
+            'consumer' => $consumer,
+            'auto_setup' => $autoSetup,
+            'stream_max_entries' => $maxEntries,
+        ], $connectionCredentials, $redisOptions, $redis);
     }
 
     public function get(): ?array
@@ -169,9 +190,15 @@ class Connection
 
         $e = null;
         try {
-            $added = $this->connection->xadd($this->stream, '*', ['message' => json_encode(
-                ['body' => $body, 'headers' => $headers]
-            )]);
+            if ($this->maxEntries) {
+                $added = $this->connection->xadd($this->stream, '*', ['message' => json_encode(
+                    ['body' => $body, 'headers' => $headers]
+                )], $this->maxEntries, true);
+            } else {
+                $added = $this->connection->xadd($this->stream, '*', ['message' => json_encode(
+                    ['body' => $body, 'headers' => $headers]
+                )]);
+            }
         } catch (\RedisException $e) {
         }